Move common types to lib.rs
Move the following generic types and traits from `report_body.rs` to
`lib.rs`:

- `EqualityVerifier`
- `GreaterThanEqualVerifier`
- `Accessor`
- `IntoVerificationError`

This move allows for other modules to make use of them.
